+++ tests/test_simple.py
@@ -1,5 +1,6 @@
 import unittest
 import pystache
+from pystache import Template
 from examples.nested_context import NestedContext
 from examples.complex_view import ComplexView
 from examples.lambdas import Lambdas
@@ -8,16 +9,15 @@ from examples.simple import Simple
 
 class TestSimple(unittest.TestCase):
 
-    def test_simple_render(self):
-        self.assertEqual('herp', pystache.Template('{{derp}}', {'derp': 'herp'}).render())
-
     def test_nested_context(self):
         view = NestedContext()
-        self.assertEquals(pystache.Template('{{#foo}}{{thing1}} and {{thing2}} and {{outer_thing}}{{/foo}}{{^foo}}Not foo!{{/foo}}', view).render(), "one and foo and two")
+        view.template = '{{#foo}}{{thing1}} and {{thing2}} and {{outer_thing}}{{/foo}}{{^foo}}Not foo!{{/foo}}'
+        self.assertEquals(view.render(), "one and foo and two")
 
     def test_looping_and_negation_context(self):
         view = ComplexView()
-        self.assertEquals(pystache.Template('{{#item}}{{header}}: {{name}} {{/item}}{{^item}} Shouldnt see me{{/item}}', view).render(), "Colors: red Colors: green Colors: blue ")
+        view.template = '{{#item}}{{header}}: {{name}} {{/item}}{{^item}} Shouldnt see me{{/item}}'
+        self.assertEquals(view.render(), "Colors: red Colors: green Colors: blue ")
 
     def test_empty_context(self):
         view = ComplexView()
@@ -25,13 +25,16 @@ class TestSimple(unittest.TestCase):
 
     def test_callables(self):
         view = Lambdas()
-        self.assertEquals(pystache.Template('{{#replace_foo_with_bar}}foo != bar. oh, it does!{{/replace_foo_with_bar}}', view).render(), 'bar != bar. oh, it does!')
+        view.template = '{{#replace_foo_with_bar}}foo != bar. oh, it does!{{/replace_foo_with_bar}}'
+        self.assertEquals(view.render(), 'bar != bar. oh, it does!')
 
     def test_rendering_partial(self):
         view = TemplatePartial()
-        self.assertEquals(pystache.Template('{{>inner_partial}}', view).render(), 'Again, Welcome!')
+        view.template = '{{>inner_partial}}'
+        self.assertEquals(view.render(), 'Again, Welcome!')
 
-        self.assertEquals(pystache.Template('{{#looping}}{{>inner_partial}} {{/looping}}', view).render(), '''Again, Welcome! Again, Welcome! Again, Welcome! ''')
+        view.template = '{{#looping}}{{>inner_partial}} {{/looping}}'
+        self.assertEquals(view.render(), '''Again, Welcome! Again, Welcome! Again, Welcome! ''')
 
     def test_non_existent_value_renders_blank(self):
         view = Simple()